+++ nodes/Hudu/resources/assets/assets.handler.ts
@@ -22,6 +22,7 @@ import {
 import type { IAssetLayoutFieldEntity } from '../asset_layout_fields/asset_layout_fields.types';
 import { isStandardField } from '../../utils/fieldTypeUtils';
 import { parseHuduApiErrorWithContext } from '../../utils/errorParser';
+import { addAssetFieldMarkdown } from '../../utils/markdown/assetFields';
 
 function toSnakeCaseFieldLabel(label: string): string {
   return label
@@ -225,7 +226,12 @@ export async function handleAssetsOperation(
 
         responseData = assets[0];
       }
-      
+
+      const includeFieldMarkdownGet = this.getNodeParameter('includeFieldMarkdown', i, false) as boolean;
+      if (includeFieldMarkdownGet) {
+        responseData = addAssetFieldMarkdown(responseData as IDataObject);
+      }
+
       debugLog('[API_RESPONSE] Get asset response', responseData);
       break;
     }
@@ -311,7 +317,12 @@ export async function handleAssetsOperation(
         returnAll,
         limit,
       );
-      
+
+      const includeFieldMarkdownGetAll = this.getNodeParameter('includeFieldMarkdown', i, false) as boolean;
+      if (includeFieldMarkdownGetAll) {
+        responseData = (responseData as IDataObject[]).map(addAssetFieldMarkdown);
+      }
+
       debugLog('[API_RESPONSE] Get all assets response');
       break;
     }